codecparsers: vc1: fix bitplanes decoding (DIFF6 or NORM6).
Fix parsing of residual bytes. This is a two-step process. First,
remaining colums of full vertical resolution (<height>) need to be
processed. Next, remaining bytes in the first row can be processed,
while taking into account the fact that we may have filled in the
first columns already. So, this is not full horizontal resolution.

Signed-off-by: Gwenole Beauchesne <gwenole.beauchesne@intel.com>