codecparsers: h264: add support for Recovery Point SEI message.
The recovery point SEI message helps a decoder in determining if the
decoding process would produce acceptable pictures for display after
the decoder initiates random access or after the encoder indicates
a broken link in the coded video sequence.

This is not used in the h264parse element, but it could help debugging.

https://bugzilla.gnome.org/show_bug.cgi?id=723380

Signed-off-by: Aurélien Zanelli <aurelien.zanelli@parrot.com>
Reviewed-by: Gwenole Beauchesne <gwenole.beauchesne@intel.com>